+++ src/main/java/com/team/cops_and_robbers/admin/application/dto/command/game/AdminGameListCommand.java
@@ -1,20 +1,15 @@
-package com.team.cops_and_robbers.admin.application.dto.command;
+package com.team.cops_and_robbers.admin.application.dto.command.game;
 
 import com.team.cops_and_robbers.admin.application.dto.SortDirection;
-import com.team.cops_and_robbers.user.domain.SocialType;
+import com.team.cops_and_robbers.game.game.domain.GameStatus;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 
-import java.time.LocalDateTime;
-
-public record AdminUserListCommand(
+public record AdminGameListCommand(
         int page,
         int size,
-        String nickname,
-        SocialType socialType,
-        LocalDateTime fromDate,
-        LocalDateTime toDate,
+        GameStatus status,
         SortDirection sortDirection
 ) {
