+++ src/main/java/com/team/cops_and_robbers/admin/application/dto/command/report/AdminUpdateReportStatusCommand.java
@@ -1,4 +1,4 @@
-package com.team.cops_and_robbers.admin.application.dto.command;
+package com.team.cops_and_robbers.admin.application.dto.command.report;
 
 import com.team.cops_and_robbers.report.domain.ReportStatus;
