+++ src/main/java/com/team/cops_and_robbers/admin/application/dto/result/dashboard/AdminDashboardResult.java
@@ -1,4 +1,4 @@
-package com.team.cops_and_robbers.admin.application.dto.result;
+package com.team.cops_and_robbers.admin.application.dto.result.dashboard;
 
 import java.util.List;
