+++ src/main/java/com/team/cops_and_robbers/admin/application/dto/result/dashboard/EndReasonDistributionResult.java
@@ -1,4 +1,4 @@
-package com.team.cops_and_robbers.admin.application.dto.result;
+package com.team.cops_and_robbers.admin.application.dto.result.dashboard;
 
 import com.team.cops_and_robbers.history.domain.GameEndReason;
 import com.team.cops_and_robbers.history.repository.EndReasonCountProjection;
