+++ src/main/java/com/team/cops_and_robbers/admin/application/dto/result/dashboard/WinRateByTeamResult.java
@@ -1,4 +1,4 @@
-package com.team.cops_and_robbers.admin.application.dto.result;
+package com.team.cops_and_robbers.admin.application.dto.result.dashboard;
 
 public record WinRateByTeamResult(
         Double policeWinRate,
