+++ src/main/java/com/team/cops_and_robbers/admin/application/dto/result/game/AdminGamePageResult.java
@@ -1,4 +1,4 @@
-package com.team.cops_and_robbers.admin.application.dto.result;
+package com.team.cops_and_robbers.admin.application.dto.result.game;
 
 import org.springframework.data.domain.Page;
