+++ src/main/java/com/team/cops_and_robbers/admin/application/dto/result/report/AdminReportPageResult.java
@@ -1,19 +1,19 @@
-package com.team.cops_and_robbers.admin.application.dto.result;
+package com.team.cops_and_robbers.admin.application.dto.result.report;
 
 import org.springframework.data.domain.Page;
 
 import java.util.List;
 
-public record AdminBugReportPageResult(
-        List<AdminBugReportResult> content,
+public record AdminReportPageResult(
+        List<AdminReportResult> content,
         Long totalElements,
         Integer totalPages,
         Integer page,
         Integer size
 ) {
 
-    public static AdminBugReportPageResult from(Page<AdminBugReportResult> page) {
-        return new AdminBugReportPageResult(
+    public static AdminReportPageResult from(Page<AdminReportResult> page) {
+        return new AdminReportPageResult(
                 page.getContent(),
                 page.getTotalElements(),
                 page.getTotalPages(),
