+++ src/main/java/com/team/cops_and_robbers/admin/presentation/AdminDashboardResolver.java
@@ -1,7 +1,7 @@
 package com.team.cops_and_robbers.admin.presentation;
 
 import com.team.cops_and_robbers.admin.application.AdminDashboardService;
-import com.team.cops_and_robbers.admin.application.dto.result.AdminDashboardResult;
+import com.team.cops_and_robbers.admin.application.dto.result.dashboard.AdminDashboardResult;
 import lombok.RequiredArgsConstructor;
 import org.springframework.graphql.data.method.annotation.QueryMapping;
 import org.springframework.stereotype.Controller;
