+++ src/main/java/com/team/cops_and_robbers/admin/presentation/AdminGameResolver.java
@@ -1,13 +1,13 @@
 package com.team.cops_and_robbers.admin.presentation;
 
 import com.team.cops_and_robbers.admin.application.AdminGameService;
-import com.team.cops_and_robbers.admin.application.dto.command.AdminGameListCommand;
-import com.team.cops_and_robbers.admin.application.dto.result.AdminGameAreaResult;
-import com.team.cops_and_robbers.admin.application.dto.result.AdminGamePageResult;
-import com.team.cops_and_robbers.admin.application.dto.result.AdminGameResult;
-import com.team.cops_and_robbers.admin.application.dto.result.AdminGameDetailResult;
-import com.team.cops_and_robbers.admin.application.dto.result.AdminParticipantResult;
 import com.team.cops_and_robbers.admin.application.dto.SortDirection;
+import com.team.cops_and_robbers.admin.application.dto.command.game.AdminGameListCommand;
+import com.team.cops_and_robbers.admin.application.dto.result.game.AdminGameAreaResult;
+import com.team.cops_and_robbers.admin.application.dto.result.game.AdminGameDetailResult;
+import com.team.cops_and_robbers.admin.application.dto.result.game.AdminGamePageResult;
+import com.team.cops_and_robbers.admin.application.dto.result.game.AdminGameResult;
+import com.team.cops_and_robbers.admin.application.dto.result.game.AdminParticipantResult;
 import com.team.cops_and_robbers.game.game.domain.GameStatus;
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
